cyan

+++ data/scripts/create-color-config.py
@@ -13,22 +13,25 @@ script_name = sys.argv[0]  # Self
 tif_name = sys.argv[1]
 out_file = sys.argv[3]
 
-if os.path.exists(out_file) and os.path.getmtime(out_file) > os.path.getmtime(tif_name):
+if (
+    "--recalc" not in sys.argv
+    and os.path.exists(out_file)
+    and os.path.getmtime(out_file) > os.path.getmtime(tif_name)
+):
     # We're here because we modified this script, not any temperature values,
     # so I'm probably just trying to change the color scheme.
     # Just rewrite the file with these new values:
     with open(out_file) as f:
-        lines =f.readlines()
+        lines = f.readlines()
     colors = [
-        '255 255 255 255', # min is white
-        f'255 {255/2} {255/2} 255', # median is in between
-        '255 0 0 255' # max is red
+        "255 0   0   255",  # min is hue angle 0° (#FF0000)
+        "255 255 255 255",  # median is saturation 0 (white)
+        "0   255 255 255",  # max is 180° (#00FFFF)
     ]
     for i, c in enumerate(colors):
-        lines[i] = lines[i].split(' ')[0]
-        lines[i] += f' {c}\n'
-    with open(out_file, 'w') as f:
-        f.write(''.join(lines))
+        lines[i] = lines[i].split(" ")[0] + f" {c}\n"
+    with open(out_file, "w") as f:
+        f.write("".join(lines))
     sys.exit(0)
 
 stem = Path(tif_name).stem
@@ -51,7 +54,7 @@ band = ds.GetRasterBand(1)
 values = np.ma.masked_equal(band.ReadAsArray(), band.GetNoDataValue())
 
 # Write out a color file based on valid values
-with open(out_file, 'w') as f:
+with open(out_file, "w") as f:
     f.write(f"{values.min()} 255 255 255 255\n")
     f.write(f"{values.mean()} 255 {255/2} {255/2} 255\n")
     f.write(f"{values.max()} 255 0 0 255\n")
